+++ Makefile
@@ -6,7 +6,7 @@
 #    By: yandry <yandry@student.42.fr>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/01/09 14:48:17 by yandry            #+#    #+#              #
-#    Updated: 2025/03/26 16:36:13 by tstephan         ###   ########.fr        #
+#    Updated: 2025/04/15 15:08:43 by tstephan         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -36,7 +36,8 @@ OBJ_PATH = obj/
 LIBFT_PATH = libft/
 
 UTILS_SRC = utils/ft_set.c utils/ft_strndup.c utils/ft_strreplace.c \
-			utils/ft_get_max.c utils/ft_strjoin_free.c utils/ft_strcmp.c
+			utils/ft_get_max.c utils/ft_strjoin_free.c utils/ft_strcmp.c \
+			utils/string_to_token.c utils/is_logical.c
 SIGNAL_SRC = signal/signal.c
 LEXING_SRC = lexing/lex.c lexing/lst_helper.c lexing/space.c \
 			  lexing/clear.c lexing/expand.c lexing/split.c \
@@ -49,7 +50,7 @@ BTREE_SRC = btree/ft_btree_clear.c btree/ft_btree_height.c \
 			btree/ft_btree_inorder.c btree/ft_btree_insert.c \
 			btree/ft_btree_new.c btree/ft_btree_postorder.c \
 			btree/ft_btree_preorder.c btree/ft_btree_size.c btree/print.c \
-			btree/ft_btree_insert_in.c
+			btree/ft_btree_insert_in.c btree/fill_tree.c
 SUBSHELL_SRC = subshell/subshell.c subshell/read.c
 PROMPT_SRC = prompt/shell.c
