+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: tstephan <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/17 16:05:54 by tstephan          #+#    #+#             */
-/*   Updated: 2025/03/25 17:44:25 by tstephan         ###   ########.fr       */
+/*   Updated: 2025/04/15 15:08:09 by tstephan         ###   ########.fr       */
 /*   Updated: 2025/03/04 16:35:23 by tstephan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
@@ -179,6 +179,8 @@ void	ft_btree_postorder(t_btree *root, void (*f)(void *));
 int		ft_btree_size(t_btree *root);
 int		ft_btree_height(t_btree *root);
 void	ft_btree_clear(t_btree **root, void (*del)(void *));
+void	ft_fill_tree(t_btree **root, t_list *pipes);
+
 char	*ft_strjoin_free(const char *s1, const char *s2, t_joinfree flag);
 
 t_leaf	*ft_create_leaf(t_node_type type, char	**args);
@@ -196,5 +198,6 @@ char	*ft_getenv(char *name);
 char	*ft_readline(t_prompt id);
 
 bool	ft_is_pipe(t_token *token);
+bool	ft_is_logical(t_token *token);
 
 #endif
